requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
idna==3.10
ifaddr==0.2.0
netdisco==3.0.0
python-nmap==0.7.1
requests==2.32.3
scapy==2.6.1
urllib3==2.4.07.0
